Implement news command on main stocks menu
