+++ gamestonk_terminal/stocks/stocks_controller.py
@@ -3,13 +3,14 @@ import os
 from typing import List
 import pandas as pd
 from prompt_toolkit.completion import NestedCompleter
+from colorama import Style
 
 from gamestonk_terminal.helper_funcs import get_flair
-
 from gamestonk_terminal.helper_funcs import b_is_stock_market_open
 from gamestonk_terminal.menu import session
 from gamestonk_terminal import feature_flags as gtff
 
+from gamestonk_terminal.common import newsapi_view
 from gamestonk_terminal.stocks.stocks_helper import load, candle, quote
 from gamestonk_terminal.stocks.discovery import disc_controller
 from gamestonk_terminal.stocks.screener import screener_controller
@@ -27,6 +28,11 @@ from gamestonk_terminal.stocks.exploratory_data_analysis import eda_controller
 from gamestonk_terminal.stocks.report import report_controller
 
 from gamestonk_terminal.options import options_controller
+
+from gamestonk_terminal.helper_funcs import (
+    check_positive,
+    parse_known_args_and_warn,
+)
 
 # pylint: disable=R1710
 
@@ -49,6 +55,7 @@ class StocksController:
         "load",
         "quote",
         "candle",
+        "news",
     ]
 
     CHOICES_MENUS = [
@@ -109,13 +116,13 @@ What do you want to do?
         else:
             help_text += "\nStock: ?\n"
 
-        help_text += f"Market {('CLOSED', 'OPEN')[b_is_stock_market_open()]}.\n"
-
-        help_text += """
-    load        load a specific stock ticker for analysis
+        help_text += f"Market {('CLOSED', 'OPEN')[b_is_stock_market_open()]}.\n\n"
+        help_text += "    load        load a specific stock ticker for analysis"
+        help_text += f"""{Style.DIM if not self.ticker else ''}
     quote       view the current price for a specific stock ticker
     candle      view a candle chart for a specific stock ticker
-
+    news        latest news of the company [News API]
+    {Style.RESET_ALL if not self.ticker else ''}
 """
         help_text += ">>  options     go into options context"
         if self.ticker:
@@ -125,11 +132,14 @@ What do you want to do?
 
 >   disc        discover trending stocks, \t e.g. map, sectors, high short interest
 >   scr         screener stocks, \t\t e.g. overview/performance, using preset filters
->   gov         government menu, \t\t house trading, contracts, corporate lobbying
->   dd          in-depth due-diligence,  \t e.g.: news, analyst, shorts, insider, sec
 >   ins         insider trading,         \t e.g.: latest penny stock buys, top officer purchases
+>   gov         government menu, \t\t e.g. house trading, contracts, corporate lobbying
+>   report      generate automatic report,   \t e.g.: dark pool, due diligence"""
+
+        help_text += f"""{Style.DIM if not self.ticker else ''}
 >   fa          fundamental analysis,    \t e.g.: income, balance, cash, earnings
 >   res         research web page,       \t e.g.: macroaxis, yahoo finance, fool
+>   dd          in-depth due-diligence,  \t e.g.: news, analyst, shorts, insider, sec
 >   ca          comparison analysis,     \t e.g.: historical, correlation, financials
 >   bt          strategy backtester,      \t e.g.: simple ema, ema cross, rsi strategies
 >   ta          technical analysis,      \t e.g.: ema, macd, rsi, adx, bbands, obv
@@ -137,8 +147,7 @@ What do you want to do?
 >   eda         exploratory data analysis,\t e.g.: decompose, cusum, residuals analysis
 >   ra          residuals analysis,      \t e.g.: model fit, qqplot, hypothesis test
 >   pred        prediction techniques,   \t e.g.: regression, arima, rnn, lstm
->   report      generate automatic report,   \t e.g.: dark pool, due diligence
-        """
+    {Style.RESET_ALL if not self.ticker else ''}"""
         print(help_text)
 
     def switch(self, an_input: str):
@@ -208,6 +217,41 @@ What do you want to do?
             self.ticker + "." + self.suffix if self.suffix else self.ticker,
             other_args,
         )
+
+    def call_news(self, other_args: List[str]):
+        """Process news command"""
+        if not self.ticker:
+            print("Use 'load <ticker>' prior to this command!", "\n")
+            return
+
+        parser = argparse.ArgumentParser(
+            add_help=False,
+            prog="news",
+            description="""
+                Prints latest news about company, including date, title and web link. [Source: News API]
+            """,
+        )
+        parser.add_argument(
+            "-n",
+            "--num",
+            action="store",
+            dest="n_num",
+            type=check_positive,
+            default=5,
+            help="Number of latest news being printed.",
+        )
+        try:
+            ns_parser = parse_known_args_and_warn(parser, other_args)
+            if not ns_parser:
+                return
+
+            newsapi_view.news(
+                term=self.ticker,
+                num=ns_parser.n_num,
+            )
+
+        except Exception as e:
+            print(e, "\n")
 
     # MENUS
     def call_disc(self, _):
